requirements.txt
numpy==1.26.4
torch==2.2.2
torchvision==0.17.2
Pillow==10.3.0
PyYAML==6.0.1
